Fixed ghibli.png, which was in the wrong place
It was in the root of the project, and React was sensibly objecting to the HomePage.jsx file trying to serve it from there. I moved it to public/images and referenced it using an explicit path in the <img> tag in that component/page. And there was much rejoicing.